Footprints in the clusters likelihood
Select a footprint by adding e.g. footprint: 'DES' to the data
section of the .yml file.

Also changed the way in which paths are given (because the
structure of the selFn directory produced by nemo is always the
same). So just two paths should be needed - one for the catalog
(cat_file) and one for the selFn directory (selfn_path).